Cleaned comments thread primitives
no issue

The focused-thread work now has shared thread graph and navigation primitives, keeping Content focused on composition while preserving the existing comments UI behavior.
